move all hand funcs to their new file. now to clean up
