+++ include/game_variables.h
@@ -27,6 +27,8 @@
 #define MAX_HANDS    4
 #define MAX_DISCARDS 4
 
+#define DEFAULT_HAND_SIZE 8
+
 /**
  * @brief A central location for all game variables.
  *
@@ -40,11 +42,15 @@ typedef struct
     s32 timer; // This might already exist in libtonc but idk so i'm just making my own
     RngInfo rng_info;
 
+    s32 hand_top; // Position of the last card in hand array, -1 when no card in hand
+    s32 hand_selections; // Number of selected Cards.
+
     // Variables visible by the player
 
     s32 round;
     s32 ante;
     s32 money;
+    s32 hand_size;
 
     // Blind variables
